requirements.txt
pcapy-ng
pyric
scapy
dpkt
paramiko
matplotlib
